+++ packages/db/src/index.ts
@@ -10,6 +10,7 @@ export {
   fakeExecutionRecordFactory,
   fakeFileRecordFactory,
   fakeGranuleRecordFactory,
+  fakeGranuleGroupRecordFactory,
   fakePdrRecordFactory,
   fakeProviderRecordFactory,
   fakeReconciliationReportRecordFactory,
@@ -40,6 +41,10 @@ export {
   PostgresCollection,
   PostgresCollectionRecord,
 } from './types/collection';
+export {
+  PostgresGranuleGroup,
+  PostgresGranuleGroupRecord,
+} from './types/granule-group';
 export {
   PostgresExecution,
   PostgresExecutionRecord,
@@ -119,6 +124,14 @@ export {
 } from './lib/collection';
 
 export {
+  DuplicateGranulesResult,
+  findActiveDuplicateGranules,
+  getNextGranuleGroupId,
+  GranuleGroupRecord,
+  GranuleInput,
+} from './lib/duplicate-detection';
+
+export {
   batchDeleteExecutionFromDatabaseByCumulusCollectionId,
   executionArnsFromGranuleIdsAndWorkflowNames,
   getApiExecutionCumulusIds,
@@ -183,6 +196,7 @@ export { CollectionPgModel } from './models/collection';
 export { ExecutionPgModel } from './models/execution';
 export { FilePgModel } from './models/file';
 export { GranulePgModel } from './models/granule';
+export { GranuleGroupsPgModel } from './models/granule-groups';
 export { GranulesExecutionsPgModel } from './models/granules-executions';
 export { PdrPgModel } from './models/pdr';
 export { ProviderPgModel } from './models/provider';